chore: keystore: move handle_ios_wal_compat to a separate module
